Use DI
