+++ src/main/java/com/andre601/formatterexpansion/formatters/number/FromTo.java
@@ -1,12 +1,21 @@
 package com.andre601.formatterexpansion.formatters.number;
 
+import com.andre601.formatterexpansion.FormatterExpansion;
 import com.andre601.formatterexpansion.formatters.IFormatter;
 import com.andre601.formatterexpansion.utils.NumberUtils;
 import com.andre601.formatterexpansion.utils.StringUtils;
 
 import java.util.Locale;
+import java.util.concurrent.TimeUnit;
 
 public class FromTo implements IFormatter{
+    
+    private final FormatterExpansion expansion;
+    
+    public FromTo(FormatterExpansion expansion){
+        this.expansion = expansion;
+    }
+    
     @Override
     public String name(){
         return null; // Never used for this special case.
@@ -23,6 +32,40 @@ public class FromTo implements IFormatter{
         if(StringUtils.isNullOrEmpty(from, to))
             return null;
     
-        return NumberUtils.convert(StringUtils.merge(1, values), from, to);
+        return convert(StringUtils.merge(1, values), from, to);
+    }
+    
+    private String convert(String number, String from, String to){
+        Long finalNumber = NumberUtils.optLong(number);
+        if(finalNumber == null)
+            return null;
+        
+        TimeUnit fromUnit = StringUtils.getTimeUnit(from);
+        TimeUnit toUnit = StringUtils.getTimeUnit(to);
+        
+        if(fromUnit == null || toUnit == null)
+            return null;
+        
+        finalNumber = toUnit.convert(finalNumber, fromUnit);
+        
+        switch(toUnit){
+            case DAYS:
+                return finalNumber + expansion.getString("time.days", "d");
+            
+            case HOURS:
+                return finalNumber + expansion.getString("time.hours", "h");
+            
+            case MINUTES:
+                return finalNumber + expansion.getString("time.minutes", "m");
+            
+            case SECONDS:
+                return finalNumber + expansion.getString("time.seconds", "s");
+            
+            case MILLISECONDS:
+                return finalNumber + expansion.getString("time.milliseconds", "ms");
+            
+            default:
+                return String.valueOf(finalNumber);
+        }
     }
 }
